display more item (track, album, artist, ...) details

+++ main/src/main/kotlin/dev/younesgouyd/apps/spotifyclient/desktop/main/data/repoes/user/CurrentUser.kt
@@ -18,6 +18,6 @@ data class CurrentUser(
 ) {
     @Serializable
     data class Followers(
-        val total: Int?
+        val total: Long?
     )
 }
